fix: modal popup component upgate: generator form store
