+++ front/src/pages/Home/components/EmailSign/index.jsx
@@ -1,7 +1,5 @@
-import { Visibility, VisibilityOff } from "@mui/icons-material";
-import { IconButton, InputAdornment, MenuItem, Select } from "@mui/material";
-import TextField from "@mui/material/TextField";
-import { Button, Modal, Upload } from "antd";
+import { MenuItem, Select } from "@mui/material";
+import { Button, Upload } from "antd";
 import { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import { FaFacebookSquare, FaLinkedin } from "react-icons/fa";
@@ -12,9 +10,8 @@ import PhoneInput from "react-phone-input-2";
 import "react-phone-input-2/lib/style.css";
 import QRCode from "react-qr-code";
 import { Link } from "react-router-dom";
-import PageHeader from "../../../../components/common/PageHeader";
 import "./_style.scss";
-import popup from "./popup.png";
+import SignUpModal from "./signUpModal";
 
 const colors = [
   { color: "000000" },
@@ -49,7 +46,6 @@ const EmailSign = () => {
   const showModal = () => {
     setIsModalOpen(true);
   };
-
   const handleCloseModal = () => {
     setIsModalOpen(false);
   };
@@ -72,11 +68,10 @@ const EmailSign = () => {
       backgroundColor: "#ffffff",
     },
   });
-
   const formValues = watch();
 
+  // Icons color default color
   const iconColor = "#1251A3";
-
   const setIconColor = () => {
     if (
       formValues.backgroundColor === "#ffffff" &&
@@ -87,6 +82,7 @@ const EmailSign = () => {
     return formValues.fontColor;
   };
 
+  // QR code generator
   const getQRCodeColors = () => {
     const fgColor = !socialUrl.qrCode
       ? "#D3D3D3"
@@ -99,12 +95,6 @@ const EmailSign = () => {
 
   const handleCategoryClick = (category) => {
     setSelectedItem(category);
-  };
-
-  const [showPassword, setShowPassword] = useState(false);
-
-  const togglePasswordVisibility = () => {
-    setShowPassword(!showPassword);
   };
 
   const handlePhoneChange = (phone) => {
@@ -190,7 +180,7 @@ const EmailSign = () => {
   return (
     <section className="signature-generator">
       <div className="container">
-<p className="font-size-28 font-weight-600 text-primary ">Generator </p>
+        <p className="font-size-28 font-weight-600 text-primary ">Generator </p>
         <div className="generator">
           <form onSubmit={handleSubmit(onSubmit)}>
             <div className="generator_banner row gap-4 mt-3">
@@ -519,131 +509,8 @@ const EmailSign = () => {
                 </Button>
               </div>
             </div>
-            <Modal
-              open={isModalOpen}
-              onClose={handleCloseModal}
-              style={{
-                display: "flex",
-                alignItems: "center",
-                justifyContent: "center",
-              }}
-              className="sign-up-modal"
-              footer={null}
-              BackdropProps={{
-                onClick: handleCloseModal,
-              }}
-            >
-              <div className="modal-content">
-                <div className="modal_head">
-                  <button
-                    type="button"
-                    aria-label="Close"
-                    className="ant-modal-close"
-                    onClick={handleCloseModal}
-                  ></button>
-                  <div className="row">
-                    <div className="col-12 col-lg-6 modal_left py-4 px-8">
-                      <div className="modal_left-header">
-                        <h3>Your Email signature is just a click away!</h3>
-                        <p>
-                          Sign up in just 30 seconds to set up and begin using
-                          your email signature free for 7 days.
-                        </p>
-                      </div>
 
-                      <form className="form-group">
-                        <div className="form-group-input">
-                          <label className="font-size-18 text-text10 font-weight-500">
-                            E-mail address
-                          </label>
-                          <TextField
-                            id="email"
-                            label="Email"
-                            variant="outlined"
-                            type="email"
-                            name="email"
-                            fullWidth
-                          />
-                        </div>
-                        <div className="form-group-input">
-                          <label className="font-size-18 text-text10 font-weight-500 ">
-                            Password
-                          </label>
-                          <TextField
-                            id="password"
-                            label="Password"
-                            variant="outlined"
-                            name="password"
-                            type={showPassword ? "text" : "password"}
-                            fullWidth
-                            InputProps={{
-                              endAdornment: (
-                                <InputAdornment position="end">
-                                  <IconButton
-                                    onClick={togglePasswordVisibility}
-                                    edge="end"
-                                  >
-                                    {showPassword ? (
-                                      <Visibility />
-                                    ) : (
-                                      <VisibilityOff />
-                                    )}
-                                  </IconButton>
-                                </InputAdornment>
-                              ),
-                            }}
-                          />
-                        </div>
-                        <div className="sign-up-buttons">
-                          <Link to="#">
-                            <button
-                              type="submit"
-                              className="btn btn-primary text-natural"
-                            >
-                              Free sign up
-                            </button>
-                          </Link>
-                          <Link to="#">
-                            <button className="google btn-google">
-                              <img
-                                src="https://t4.ftcdn.net/jpg/03/08/54/37/360_F_308543787_DmPo1IELtKY9hG8E8GlW8KHEsRC7JiDN.jpg"
-                                alt="google sign up"
-                              />
-                              Continue with Google
-                            </button>
-                          </Link>
-                        </div>
-                      </form>
-
-                      <div className="modal_left-footer">
-                        <div className="login">
-                          <span>Already have an account? - </span>
-                          <Link to="/login">Login</Link>
-                        </div>
-                        <div className="free-trial">
-                          <div className="privacy-policy">
-                            <p>
-                              By signing up, you agree to the
-                              <Link to="/"> Terms And Condition</Link>
-                            </p>
-                            <p>
-                              and
-                              <Link to="/"> Privacy Policy</Link>
-                            </p>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-
-                    <div className="col-12 col-lg-6 modal_right d-none">
-                      <div className="modal_right_img">
-                        <img src={popup} alt="" />
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </Modal>
+            <SignUpModal isOpen={isModalOpen} handleClose={handleCloseModal} />
           </form>
         </div>
       </div>
